move doc serializer to own file
